refactor: Prepare uplink for mobile
